Avoid unintended double-slide switches
It appears that scroll wheel events keep getting processed long after the user actually removes their fingers from the trackpad, with the result that if I swipe for 800ms, even if the swipe duration is 1000ms, it will usually end up triggering a second swipe after the first one finishes.

This fixes that by making sure the *speed* of the scroll wheel has just spiked (and wasn't already spiking earlier), so that we only get one slide change per swipe motion on a trackpad.

Feel free to play around with this! Thanks a lot for this library—we're using it in a new repository at Tictail, and we subclassed `Deck` with this change in order to get the behavior we wanted. It would be nice if the change were made upstream, though, so we could get rid of our child class.